move extract bit
